backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
opencv-python-headless==4.8.1.78
ffmpeg-python==0.2.0
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0
email-validator==2.1.0